moved yoast out of header directory
