+++ src/client/apps/edit/components/edit_container.jsx
@@ -13,9 +13,9 @@ import EditAdmin from "./admin"
 import EditContent from "./content"
 import EditDisplay from "./display"
 import EditError from "./error"
-import EditHeader from "./header/index.tsx"
+import EditHeader from "./header"
 import { MessageModal } from "./message"
-import Yoast from "./header/yoast"
+import Yoast from "./yoast"
 import styled from "styled-components"
 
 const INACTIVITY_TIMEOUT = 600 * 1000
